[iCub] minor refactoring
